Move info helpers to its own class
